feat: stopped double import, bug fix on app.yaml
Put all the db interaction based stuff into the app.py file to avoid double import problem.

Changed connector to be the correct one

Type: refactor
